requirements.txt
langchain>=1.2.10
langchain-core>=1.2.13
langchain-google-genai>=4.2.0
langchain-ollama>=1.0.1
langgraph>=1.0.8
ollama>=0.6.1
google-genai>=1.64.0
python-dotenv>=1.0.0
aiohttp>=3.13.3
httpx>=0.28.1
